+++ src/Symfony/Component/Mailer/Envelope.php
@@ -44,10 +44,6 @@ class Envelope
 
     public function setSender(Address $sender): void
     {
-        // to ensure deliverability of bounce emails independent of UTF-8 capabilities of SMTP servers
-        if (!preg_match('/^[^@\x80-\xFF]++@/', $sender->getAddress())) {
-            throw new InvalidArgumentException(\sprintf('Invalid sender "%s": non-ASCII characters not supported in local-part of email.', $sender->getAddress()));
-        }
         $this->sender = $sender;
     }
 
@@ -87,13 +83,28 @@ class Envelope
     }
 
     /**
+
+      * Returns true if any address' localpart contains at least one
+      * non-ASCII character, and false if all addresses have all-ASCII
+      * localparts.
+      *
+      * This helps to decide whether to the SMTPUTF8 extensions (RFC
+      * 6530 and following) for any given message.
+      *
+      * The SMTPUTF8 extension is strictly required if any address
+      * contains a non-ASCII character in its localpart. If non-ASCII
+      * is only used in domains (e.g. horst@freiherr-von-mühlhausen.de)
+      * then it is possible to to send the message using IDN encoding
+      * instead of SMTPUTF8. The most common software will display the
+      * message as intended.
+      *
       * @return bool
       */
     public function anyAddressHasUnicodeLocalpart(): bool
     {
-        if($this->sender->hasUnicodeLocalpart())
+        if($this->getSender()->hasUnicodeLocalpart())
             return true;
-        foreach($this->recipients as $r)
+        foreach($this->getRecipients() as $r)
             if($r->hasUnicodeLocalpart())
                 return true;
         return false;