+++ src/Symfony/Component/Mailer/Tests/Transport/Smtp/EsmtpTransportTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Mailer\Tests\Transport\Smtp;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Mailer\Exception\InvalidArgumentException;
 use Symfony\Component\Mailer\Exception\TransportException;
 use Symfony\Component\Mailer\Transport\Smtp\Auth\CramMd5Authenticator;
 use Symfony\Component\Mailer\Transport\Smtp\Auth\LoginAuthenticator;
@@ -60,6 +61,37 @@ class EsmtpTransportTest extends TestCase
 
         $this->assertContains("MAIL FROM:<sender@example.org> RET=HDRS\r\n", $stream->getCommands());
         $this->assertContains("RCPT TO:<recipient@example.org> NOTIFY=FAILURE\r\n", $stream->getCommands());
+    }
+
+    public function testSmtputf8()
+    {
+        $stream = new DummyStream();
+        $transport = new Smtputf8EsmtpTransport(stream: $stream);
+
+        $message = new Email();
+        $message->from('info@dømi.fo');
+        $message->addTo('dømi@dømi.fo');
+        $message->text('.');
+
+        $transport->send($message);
+
+        $this->assertContains("MAIL FROM:<info@xn--dmi-0na.fo> SMTPUTF8\r\n", $stream->getCommands());
+        $this->assertContains("RCPT TO:<dømi@xn--dmi-0na.fo>\r\n", $stream->getCommands());
+    }
+
+    public function testMissingSmtputf8()
+    {
+        $stream = new DummyStream();
+        $transport = new EsmtpTransport(stream: $stream);
+
+        $message = new Email();
+        $message->from('info@dømi.fo');
+        $message->addTo('dømi@dømi.fo');
+        $message->text('.');
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Invalid addresses: non-ASCII characters not supported in local-part of email.');
+        $transport->send($message);
     }
 
     public function testConstructorWithDefaultAuthenticators()
@@ -265,6 +297,20 @@ class CustomEsmtpTransport extends EsmtpTransport
 
         if (str_starts_with($command, 'EHLO ')) {
             $response .= "250 DSN\r\n";
+        }
+
+        return $response;
+    }
+}
+
+class Smtputf8EsmtpTransport extends EsmtpTransport
+{
+    public function executeCommand(string $command, array $codes): string
+    {
+        $response = parent::executeCommand($command, $codes);
+
+        if (str_starts_with($command, 'EHLO ')) {
+            $response .= "250 SMTPUTF8\r\n";
         }
 
         return $response;