+++ src/Symfony/Component/Mailer/Transport/Smtp/EsmtpTransport.php
@@ -195,6 +195,11 @@ class EsmtpTransport extends SmtpTransport
         return $capabilities;
     }
 
+    protected function serverSupportsSmtputf8(): bool
+    {
+        return \array_key_exists('SMTPUTF8', $this->capabilities);
+    }
+
     private function handleAuth(array $modes): void
     {
         if (!$this->username) {